requirements.txt
cloudpickle==3.0.0
dill
Flask==3.0.3
keras==3.5.0
numpy==1.23.5
opencv-python-headless
pandas==2.2.2
Pillow==10.4.0
Requests==2.32.3
scikit-image==0.24.0
scikit_learn==1.5.2
tensorflow==2.17.0
torch==2.2.1
torchvision==0.17.1
transformers==4.44.2
xgboost==2.1.1